feat: add upsert_alliance service method and implement on handle_job method

+++ src/server/model/worker.rs
@@ -2,5 +2,5 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum WorkerJob {
-    UpdateAlliance { alliance_id: i32 },
+    UpdateAlliance { alliance_id: i64 },
 }
